refactor: address review feedback on API resource timing
- Move the timing-breakdown helper out of plugins/ (it is not a plugin)
  to src/network-timings.ts
- Make calcTotalTime a pure function that returns total instead of
  mutating timings; callers assign the result
- Restore the original TTFB / responseEnd comments in network.ts
